refactor: change 'isEqualReportPeriodStart' to 'isEqualMonthAndYear'
